pyproject.toml
[tool.poetry]
name = "myhh"
version = "0.1.0"
description = ""
authors = ["Zag207 <danilka.stepanov04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
psycopg = {extras = ["binary"], version = "^3.2.1"}
asyncpg = "^0.29.0"

[tool.poetry.dev-dependencies]
ruff = "^0.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
